Move interfaces to package restic/types
